Refactor: move StageMenu logic in Header, replace action as function prop
